api/requirements.txt
attrdict==2.0.0
urlpath==1.1.2
pathlib2==2.1.0
ruamel.yaml==0.12.15
requests==2.11.1
aiohttp==1.2.0
Flask==0.10.1
Jinja2==2.7.3
Werkzeug==0.10.4
gunicorn==19.3.0
cryptography
packaging==16.8
